# webapp/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TerrAI Spatial Intelligence API",
    "description": "Read-only file-backed APIs for the TerrAI commercial exhibition prototype.",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/health": {
      "get": {
        "tags": [
          "reliability"
        ],
        "summary": "Health",
        "description": "Return API and packaged-dataset readiness.",
        "operationId": "health_api_v1_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Health Api V1 Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/catalog": {
      "get": {
        "tags": [
          "reliability"
        ],
        "summary": "Catalog",
        "description": "Return the auditable file-backed dataset catalog.",
        "operationId": "catalog_api_v1_catalog_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Catalog Api V1 Catalog Get"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bootstrap": {
      "get": {
        "tags": [
          "exhibition"
        ],
        "summary": "Bootstrap",
        "description": "Return the complete local exhibition payload through one stable contract.",
        "operationId": "bootstrap_api_v1_bootstrap_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Bootstrap Api V1 Bootstrap Get"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets/{key}": {
      "get": {
        "tags": [
          "data"
        ],
        "summary": "Dataset",
        "description": "Return one JSON or GeoJSON dataset by stable public key.",
        "operationId": "dataset_api_v1_datasets__key__get",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Dataset Api V1 Datasets  Key  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/features/{key}": {
      "get": {
        "tags": [
          "query"
        ],
        "summary": "Features",
        "description": "Filter, sort and spatially window a GeoJSON FeatureCollection.",
        "operationId": "features_api_v1_features__key__get",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          },
          {
            "name": "where",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Where"
            }
          },
          {
            "name": "equals",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Equals"
            }
          },
          {
            "name": "minimum",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Minimum"
            }
          },
          {
            "name": "maximum",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Maximum"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sort"
            }
          },
          {
            "name": "descending",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Descending"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5000,
              "minimum": 1,
              "default": 5000,
              "title": "Limit"
            }
          },
          {
            "name": "bbox",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "number"
                  },
                  "minItems": 4,
                  "maxItems": 4
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bbox"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Features Api V1 Features  Key  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recommendations/{analysis}": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "Recommendations",
        "description": "Return a server-ranked action queue for one exhibition analysis.",
        "operationId": "recommendations_api_v1_recommendations__analysis__get",
        "parameters": [
          {
            "name": "analysis",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Recommendations Api V1 Recommendations  Analysis  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scenes": {
      "get": {
        "tags": [
          "scenes"
        ],
        "summary": "Scenes",
        "description": "Return the renderer-neutral underground scene catalog.",
        "operationId": "scenes_api_v1_scenes_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Scenes Api V1 Scenes Get"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scenes/{scene_id}": {
      "get": {
        "tags": [
          "scenes"
        ],
        "summary": "Scene Bundle",
        "description": "Return one catalogued scene with its local-frame handoff.",
        "operationId": "scene_bundle_api_v1_scenes__scene_id__get",
        "parameters": [
          {
            "name": "scene_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scene Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Scene Bundle Api V1 Scenes  Scene Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}